+++ Tests/LinuxMain.swift
@@ -392,6 +392,7 @@ extension RulesTests {
         ("testCyclomaticComplexity", testCyclomaticComplexity),
         ("testDiscardedNotificationCenterObserver", testDiscardedNotificationCenterObserver),
         ("testDiscouragedObjectLiteral", testDiscouragedObjectLiteral),
+        ("testDiscouragedOptionalBoolean", testDiscouragedOptionalBoolean),
         ("testDynamicInline", testDynamicInline),
         ("testEmptyCount", testEmptyCount),
         ("testEmptyEnumArguments", testEmptyEnumArguments),