+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 
 #### Enhancements
 
-* None.
+* Adds `discouraged_optional_boolean` opt-in rules to discourage
+  the use of optional booleans.  
+  [Ornithologist Coder](https://github.com/ornithocoder)
+  [#2011](https://github.com/realm/SwiftLint/issues/2011)
 
 #### Bug Fixes